add hashedstatewriter impl
