+++ crates/storage/provider/src/bundle_state/mod.rs
@@ -1,11 +1,9 @@
 //! Bundle state module.
 //! This module contains all the logic related to bundle state.
 mod bundle_state_with_receipts;
-mod hashed_state_changes;
 mod state_reverts;
 
 pub use bundle_state_with_receipts::{
     AccountRevertInit, BundleStateInit, BundleStateWithReceipts, OriginalValuesKnown, RevertsInit,
 };
-pub use hashed_state_changes::HashedStateChanges;
 pub use state_reverts::StorageRevertsIter;
